Moved the code files from their own seperate packages to just all be inside the Aircraft package, should hopefully work now :), i also had to tweak some of the code a little bit because certain things contained like .domain in them and what not, but that had to be changed because it doesnt exist anymore but all should work well now! :)
